requirements.txt
groq==0.18.0
mediapipe==0.10.21
numpy==2.2.3
opencv_python_headless==4.10.0.84
python-dotenv==1.0.1
